Remove iab_model_output model artifacts
Delete generated model output files under iab_model_output: config.json, tokenizer.json, tokenizer_config.json, and train_metrics.json. These files contained the DistilBERT model configuration, tokenizer data, and training metrics and are being removed to clean up generated artifacts and reduce repository size.

+++ iab_retrieval.py
@@ -191,7 +191,7 @@ class LocalTextEmbedder:
             else:
                 mask = inputs["attention_mask"].unsqueeze(-1)
                 pooled = (hidden * mask).sum(dim=1) / mask.sum(dim=1).clamp(min=1)
-            rows.append(F.normalize(pooled, p=2, dim=1).cpu())
+            rows.append(F.normalize(pooled.float(), p=2, dim=1).cpu())
         return torch.cat(rows, dim=0)
